Allow for more than 255 LEDs in a display.

+++ src/display_properties.c
@@ -38,7 +38,7 @@ void init_display_properties() {
   dp_info_range.end = eeprom_read_byte(&DP_LED_INFORMATION.count);
 }
 
-uint8_t get_led_count() {
+uint16_t get_led_count() {
   return dp_info_range.end;
 }
